local-airflow/requirements.txt
mlflow==3.1.1
numpy>=2.3.1
optuna>=4.4.0
pandas>=2.3.0
scikit-learn>=1.7.0
xgboost>=2.0.0
google-cloud-storage>=2.10.0
psycopg2-binary>=2.9.0
evidently>=0.7.10
pytest>=8.0.0
python-dotenv>=1.0.0
